+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Add `Container::changes` method
 * Add `Container::logs` method
 * Add `Container::stats`, `Container::stats_stream`, `Containers::stats` and `Containers::stats_stream` methods
+* Add `Container::top` and `Container::top_stream` methods